# api/swagger/openapi.json
{
  "components": {
    "schemas": {
...
      "ConnectionConfig": {
        "properties": {
          "created_at": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "is_default": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "SuccessResponse": {
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "NATS Horizon Project",
      "url": "http://github.com/nats-horizon"
    },
    "description": "Comprehensive API for monitoring and managing NATS JetStream instances.",
    "license": {
      "name": "Apache-2.0",
      "url": "https://opensource.org/licenses/Apache-2.0"
    },
    "termsOfService": "http://swagger.io/terms/",
    "title": "NATS Horizon API",
    "version": "1.0"
  },
...
  "openapi": "3.1.0",
  "paths": {
...
    "/tenancy/connections": {
      "get": {
        "description": "Returns all configured multi-tenancy NATS connections",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "connections list"
          }
        },
        "summary": "List tenancy connections",
        "tags": [
          "tenancy"
        ]
      },
      "post": {
        "description": "Creates a new multi-tenancy NATS connection configuration",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionConfig"
              }
            }
          },
          "description": "Connection configuration",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionConfig"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          }
        },
        "summary": "Create a tenancy connection",
        "tags": [
          "tenancy"
        ]
      }
    },
    "/tenancy/connections/test": {
      "post": {
        "description": "Attempts to connect to a NATS URL and reports connectivity and latency",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "description": "Connection test request",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "connection test result"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          }
        },
        "summary": "Test a connection",
        "tags": [
          "tenancy"
        ]
      }
    },
    "/tenancy/connections/{id}": {
      "delete": {
        "description": "Deletes a multi-tenancy NATS connection configuration (the default connection cannot be deleted)",
        "parameters": [
          {
            "description": "Connection ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "summary": "Delete a tenancy connection",
        "tags": [
          "tenancy"
        ]
      },
      "get": {
        "description": "Returns a single configured NATS connection by ID",
        "parameters": [
          {
            "description": "Connection ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionConfig"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "summary": "Get a tenancy connection",
        "tags": [
          "tenancy"
        ]
      },
      "put": {
        "description": "Updates an existing multi-tenancy NATS connection configuration",
        "parameters": [
          {
            "description": "Connection ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionConfig"
              }
            }
          },
          "description": "Connection configuration",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionConfig"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "summary": "Update a tenancy connection",
        "tags": [
          "tenancy"
        ]
      }
    },
    "/tenancy/connections/{id}/default": {
      "put": {
        "description": "Marks a connection as the default NATS connection",
        "parameters": [
          {
            "description": "Connection ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionConfig"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "summary": "Set default connection",
        "tags": [
          "tenancy"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api"
    }
  ]
}